move CondaPackageRequirement to mlem.contrib.venv
