+++ setup.py
@@ -196,7 +196,7 @@ setup_args = dict(  # noqa: C408
             "builder.requirements = mlem.contrib.requirements:RequirementsBuilder",
             "builder.venv = mlem.contrib.venv:VenvBuilder",
             "builder.conda = mlem.contrib.venv:CondaBuilder",
-            "requirement.conda = mlem.contrib.requirements:CondaPackageRequirement",
+            "requirement.conda = mlem.contrib.venv:CondaPackageRequirement",
             "client.rmq = mlem.contrib.rabbitmq:RabbitMQClient",
             "server.rmq = mlem.contrib.rabbitmq:RabbitMQServer",
             "docker_registry.ecr = mlem.contrib.sagemaker.build:ECRegistry",
